requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-htmx
build==1.2.1
    # via pip-tools
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
cryptography==43.0.3
    # via pyjwt
distlib==0.3.8
    # via virtualenv
django==5.0.7
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-htmx
django-allauth[socialaccount]==65.2.0
    # via -r requirements.in
django-autoslug==1.9.9
    # via -r requirements.in
django-htmx==1.18.0
    # via -r requirements.in
django-utils-six==2.0
    # via -r requirements.in
filelock==3.15.4
    # via virtualenv
identify==2.6.0
    # via pre-commit
idna==3.10
    # via requests
nodeenv==1.9.1
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via virtualenv
pre-commit==3.7.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via django-allauth
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
sqlparse==0.5.1
    # via django
urllib3==2.2.3
    # via requests
virtualenv==20.26.3
    # via pre-commit
wheel==0.43.0
    # via pip-tools
whitenoise==6.7.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
